requirements.txt
brotlipy==0.7.0
certifi==2022.5.18.1
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056590078/work
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588893/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz==0.11.0
dask @ file:///opt/conda/conda-bld/dask-core_1653082102769/work
fsspec @ file:///opt/conda/conda-bld/fsspec_1652802105984/work
glfw==2.2.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imagecodecs @ file:///tmp/build/80754af9/imagecodecs_1635529108216/work
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
imageio-ffmpeg==0.4.3
imgui==1.3.0
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1613407855456/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
networkx @ file:///opt/conda/conda-bld/networkx_1647437648384/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.6.0.66
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.1.5
partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
Pillow==8.4.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydicom==2.3.0
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.1
pylibjpeg-openjpeg==1.2.1
PyOpenGL==3.1.5
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pyspng==0.1.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1648710015787/work
PyYAML==6.0
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
scikit-image @ file:///tmp/build/80754af9/scikit-image_1648214171611/work
scipy @ file:///tmp/build/80754af9/scipy_1641555004408/work
SimpleITK==2.2.0rc3
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
tifffile @ file:///tmp/build/80754af9/tifffile_1627275862826/work
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
torch==1.9.1
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
tqdm @ file:///tmp/build/80754af9/tqdm_1631818572807/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
